# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Financiamento Veículos",
        "description": "API para gerenciamento do sistema de financiamento de veículos",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/clientes": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Lista todos os clientes",
                "operationId": "listarClientes",
                "responses": {
                    "200": {
                        "description": "Lista retornada com sucesso"
                    }
                }
            },
            "post": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Cadastrar cliente",
                "operationId": "criarCliente",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "cpf"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Edson Belém"
                                    },
                                    "cpf": {
                                        "type": "string",
                                        "example": "12345678901"
                                    },
                                    "data_nascimento": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1990-05-10"
                                    },
                                    "telefone": {
                                        "type": "string",
                                        "example": "(24)99999-9999"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "edson@gmail.com"
                                    },
                                    "cep": {
                                        "type": "string",
                                        "example": "27510000"
                                    },
                                    "logradouro": {
                                        "type": "string",
                                        "example": "Rua das Flores"
                                    },
                                    "numero": {
                                        "type": "string",
                                        "example": "123"
                                    },
                                    "bairro": {
                                        "type": "string",
                                        "example": "Centro"
                                    },
                                    "cidade": {
                                        "type": "string",
                                        "example": "Resende"
                                    },
                                    "uf": {
                                        "type": "string",
                                        "example": "RJ"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cliente criado com sucesso"
                    }
                }
            }
        },
        "/api/clientes/{id}": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Buscar cliente por ID",
                "operationId": "buscarCliente",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do cliente",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cliente encontrado"
                    },
                    "404": {
                        "description": "Cliente não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Atualizar cliente",
                "operationId": "atualizarCliente",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do cliente",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Edson Belém"
                                    },
                                    "cpf": {
                                        "type": "string",
                                        "example": "12345678901"
                                    },
                                    "data_nascimento": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1990-05-10"
                                    },
                                    "telefone": {
                                        "type": "string",
                                        "example": "(24)99999-9999"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "edson@gmail.com"
                                    },
                                    "cep": {
                                        "type": "string",
                                        "example": "27510000"
                                    },
                                    "logradouro": {
                                        "type": "string",
                                        "example": "Rua das Flores"
                                    },
                                    "numero": {
                                        "type": "string",
                                        "example": "123"
                                    },
                                    "bairro": {
                                        "type": "string",
                                        "example": "Centro"
                                    },
                                    "cidade": {
                                        "type": "string",
                                        "example": "Resende"
                                    },
                                    "uf": {
                                        "type": "string",
                                        "example": "RJ"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cliente atualizado com sucesso"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Excluir cliente",
                "operationId": "excluirCliente",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do cliente",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cliente removido com sucesso"
                    }
                }
            }
        },
...
    },
...
}